fix bug, cleaning code

+++ projects/client/src/pages/user/ProductDetail.jsx
@@ -8,6 +8,7 @@ import { useSelector, useDispatch } from "react-redux";
 import { countItem } from "../../redux/cartSlice";
 import { CheckIcon } from '@heroicons/react/20/solid'
 import { MinusCircleIcon, PlusCircleIcon } from "@heroicons/react/24/outline";
+import ProductInformation from "../../component/ProductInformation";
 
 const ProductDetail = () => {
     const dispatch = useDispatch()
@@ -43,8 +44,7 @@ const ProductDetail = () => {
               'headers': {
                   'Authorization': `Bearer ${token}`
               }
-          });
-  
+          }); 
           dispatch(
             countItem({
               count: response.data.data,
@@ -58,10 +58,8 @@ const ProductDetail = () => {
 
     async function addToCart(){
         try{
-          // const data = {quantity: productQty}
             if(carts.length>0){
                 let deleteCart = false;
-
                 for(let item of carts){
                     if(item.id_branch!=branchId){
                         deleteCart = true;
@@ -71,37 +69,19 @@ const ProductDetail = () => {
 
                 if(deleteCart){
                     if (window.confirm("You have items from other branch on your cart. Delete current cart?") == true) {
-                        const deleteCart = await api.delete(`cart`, {
-                            'headers': {
-                                'Authorization': `Bearer ${token}`
-                            }
-                        });
+                        const deleteCart = await api.delete(`cart`, {'headers': {'Authorization': `Bearer ${token}`}});
                         toast.success(deleteCart.data.message);
-                        const response = await api.post(`cart/${id}`, {quantity: productQty, bonusQty}, {
-                            'headers': {
-                                'Authorization': `Bearer ${token}`
-                            }
-                        });
-
-                        countCart()
+                        const response = await api.post(`cart/${id}`, {quantity: productQty, bonusQty}, {'headers': {'Authorization': `Bearer ${token}`}});
                         toast.success(response.data.message);
+                        window.location.href = `/product/${id}`
                     }
                 }else{
-                    const response = await api.post(`cart/${id}`, {quantity: productQty, bonusQty}, {
-                        'headers': {
-                            'Authorization': `Bearer ${token}`
-                        }
-                    });
-
+                    const response = await api.post(`cart/${id}`, {quantity: productQty, bonusQty}, {'headers': {'Authorization': `Bearer ${token}`}});
                     countCart()
                     toast.success(response.data.message);
                 }
             }else{
-                const response = await api.post(`cart/${id}`, {quantity: productQty, bonusQty}, {
-                    'headers': {
-                        'Authorization': `Bearer ${token}`
-                    }
-                });
+                const response = await api.post(`cart/${id}`, {quantity: productQty, bonusQty}, {'headers': {'Authorization': `Bearer ${token}`}});
                 countCart()
                 toast.success(response.data.message);
             }
@@ -168,87 +148,13 @@ const ProductDetail = () => {
         setProductQty(qty)
       )
     }
-    function formatIDR(price) {
-      let idr = Math.ceil(price).toLocaleString("id-ID");
-      return `Rp ${idr}`;
-    }
 
     return (
       <>
         <NavBar />
         <div className="bg-white">
           <div className="mx-auto max-w-2xl py-8 px-4 sm:py-24 sm:px-6 lg:grid lg:max-w-7xl lg:grid-cols-2 lg:gap-x-10 lg:py-10 lg:px-8">
-            {/* Product image */}
-            <div className=" lg:col-start-1 lg:row-start-1 lg:row-span-2 lg:mt-0 lg:self-center">
-              <div className="aspect-w-1 aspect-h-1 overflow-hidden rounded-lg">
-                <img
-                  src={product.Product?.product_image}
-                  alt={product.Product?.product_name}
-                  className="h-full w-full object-cover object-center"
-                />
-              </div>
-            </div>
-
-            {/* Product details */}
-            <div className="lg:max-w-lg lg:self-end lg:col-start-2 lg:self-start">
-              <div className="mt-4">
-                <h1 className="text-2xl font-bold tracking-tight text-gray-900 sm:text-4xl lg:text-3xl">
-                  {product.Product?.product_name}
-                </h1>
-              </div>
-
-              <section aria-labelledby="information-heading" className="mt-4">
-                <h2 id="information-heading" className="sr-only">
-                  Product information
-                </h2>
-                <div className="flex items-center">
-                  <div className="text-xl font-bold text-green-500 sm:text-2xl mr-5 pb-1">
-                    {formatIDR(product?.discounted_price)}
-                  </div>
-
-                  {product.Discounts && (
-                    <div className="text-sm bg-orange-400 text-white font-semibold px-2 rounded-full">
-                      {product.Discounts?.[0]?.discount_type === "buy one get one"
-                        ? product.Discounts?.[0]?.discount_type
-                        : product.Discounts?.[0]?.discount_type === "percentage"
-                        ? `${product.Discounts?.[0]?.discount_value}%`
-                        : product.Discounts?.[0]?.discount_type === "amount"
-                        ? `${product.Discounts?.[0]?.discount_value} off`
-                        : null}
-                    </div>
-                  )}
-                  {product.Discounts &&
-                    (product.Discounts?.[0]?.discount_type === "amount" ||
-                    product.Discounts?.[0]?.discount_type === "percentage") && (
-                      <div className="ml-3 mb-0.5 text-lg text-gray-400 line-through">
-                        {formatIDR(product.Product?.product_price)}
-                      </div>
-                    )}
-                </div>
-
-                <div className="mt-4 space-y-6">
-                  <p className="text-lg text-gray-500">
-                    {product.Product?.product_description}
-                  </p>
-                </div>
-
-                <div className="mt-6 flex items-center">
-                  <CheckIcon
-                    className="h-5 w-5 flex-shrink-0 text-green-500"
-                    aria-hidden="true"
-                  />
-                  <p className="ml-2 text-md text-gray-500">
-                    Weight : {product.Product?.weight} gram
-                  </p>
-                </div>
-                <div className="mt-2 flex items-center">
-                      <CheckIcon className="h-5 w-5 flex-shrink-0 text-green-500" aria-hidden="true" />
-                      <p className="ml-2 text-md text-gray-500">
-                        In stock : {product.stock}</p>
-                </div>
-              </section>
-            </div>
-
+            <ProductInformation product={product}/>
             <div className="mt-4 lg:col-start-2 lg:row-start-2 lg:max-w-lg lg:self-start">
               <section aria-labelledby="options-heading">
                   <div className="mt-10">
